+++ namecheap/domains_dns_set_hosts.go
@@ -129,11 +129,6 @@ func (dds *DomainsDNSService) SetHostsWithContext(ctx context.Context, args *Dom
 	if err != nil {
 		return nil, err
 	}
-	if response.Errors != nil && len(*response.Errors) > 0 {
-		apiErr := (*response.Errors)[0]
-		return nil, fmt.Errorf("%s (%s)", *apiErr.Message, *apiErr.Number)
-	}
-
 	return response.CommandResponse, nil
 }