+++ namecheap/domains_ns_get_info.go
@@ -3,7 +3,6 @@ package namecheap
 import (
 	"context"
 	"encoding/xml"
-	"fmt"
 )
 
 type NameserversGetInfoResponse struct {
@@ -42,11 +41,6 @@ func (s *DomainsNSService) GetInfoWithContext(ctx context.Context, sld, tld, nam
 	_, err := s.client.DoXMLWithContext(ctx, params, &response)
 	if err != nil {
 		return nil, err
-	}
-
-	if response.Errors != nil && len(*response.Errors) > 0 {
-		apiErr := (*response.Errors)[0]
-		return nil, fmt.Errorf("%s (%s)", *apiErr.Message, *apiErr.Number)
 	}
 
 	return response.CommandResponse, nil